Feature/order

+++ src/main/java/com/pillow/pillow/product/dto/ProductResponseDTO.java
@@ -3,6 +3,7 @@ package com.pillow.pillow.product.dto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -13,7 +14,7 @@ public class ProductResponseDTO {
     private UUID id;
     private String title;
     private String description;
-    private Integer price;
+    private BigDecimal price;
     private List<String> images;
 
 }
